fixed gitlab runner build failure

+++ Makefile
@@ -3,4 +3,4 @@ mp:
 standard:
 	g++ -fopenmp -std=c++11 -static-libgcc -static-libstdc++ main.cpp -o main
 clean:
-	rm main
+	[ -e main ] && rm main
